pyproject.toml
[project]
name = "aegis"
version = "0.1.0"
description = "Aegis — AI content discovery and curation bot for Discord"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3",
    "google-api-python-client>=2.100",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "openai>=1.50",
    "APScheduler>=3.10",
    "aiosqlite>=0.20",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.5",
]

[project.scripts]
aegis = "src.main:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
